@@ -1,101 +1,40 @@
 {
-  "name": "reactjs",
+  "name": "reacts",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "antd-mobile": "^1.6.2",
-    "autoprefixer": "^7.1.2",
-    "babel-core": "^6.25.0",
-    "babel-eslint": "^7.2.3",
-    "babel-jest": "^20.0.3",
-    "babel-loader": "^7.1.1",
-    "babel-preset-react-app": "^3.0.2",
-    "babel-runtime": "^6.23.0",
-    "case-sensitive-paths-webpack-plugin": "^2.1.1",
-    "chalk": "^1.1.3",
-    "copy-to-clipboard": "^3.0.8",
-    "css-loader": "^0.28.4",
-    "dotenv": "^4.0.0",
-    "eslint": "^4.4.1",
-    "eslint-config-react-app": "^2.0.0",
-    "eslint-loader": "^1.9.0",
-    "eslint-plugin-flowtype": "^2.35.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-react": "^7.1.0",
-    "extract-text-webpack-plugin": "^3.0.0",
-    "file-loader": "^0.11.2",
-    "fs-extra": "^3.0.1",
-    "html-webpack-plugin": "^2.29.0",
-    "jest": "^20.0.4",
-    "object-assign": "^4.1.1",
-    "postcss-flexbugs-fixes": "^3.2.0",
-    "postcss-loader": "^2.0.6",
-    "promise": "^8.0.1",
-    "react": "^15.6.1",
-    "react-dev-utils": "^3.1.1",
-    "react-dom": "^15.6.1",
-    "react-error-overlay": "^1.0.10",
-    "react-router-dom": "^4.1.2",
-    "style-loader": "^0.18.2",
-    "sw-precache-webpack-plugin": "^0.11.4",
-    "url-loader": "^0.5.9",
-    "webpack": "^3.5.1",
-    "webpack-dev-server": "^2.7.1",
-    "webpack-manifest-plugin": "^1.2.1",
-    "whatwg-fetch": "^2.0.3"
+    "antd-mobile": "^2.3.1",
+    "babel-plugin-import": "^1.12.0",
+    "copy-to-clipboard": "^3.2.0",
+    "customize-cra": "^0.5.0",
+    "less": "^3.10.1",
+    "less-loader": "^5.0.0",
+    "moment": "^2.24.0",
+    "react": "^16.9.0",
+    "react-app-rewired": "^2.1.3",
+    "react-dom": "^16.9.0",
+    "react-router-dom": "^5.0.1",
+    "react-scripts": "3.1.1"
   },
   "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js --env=jsdom"
-  },
-  "devDependencies": {
-    "babel-plugin-import": "^1.3.1",
-    "less": "^2.7.2",
-    "less-loader": "^4.0.5",
-    "postcss-pxtorem": "^4.0.1",
-    "svg-sprite-loader": "^0.3.1"
-  },
-  "jest": {
-    "collectCoverageFrom": [
-      "src/**/*.{js,jsx}"
-    ],
-    "setupFiles": [
-      "<rootDir>/config/polyfills.js"
-    ],
-    "testMatch": [
-      "<rootDir>/src/**/__tests__/**/*.js?(x)",
-      "<rootDir>/src/**/?(*.)(spec|test).js?(x)"
-    ],
-    "testEnvironment": "node",
-    "testURL": "http://localhost",
-    "transform": {
-      "^.+\\.(js|jsx)$": "<rootDir>/node_modules/babel-jest",
-      "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-      "^(?!.*\\.(js|jsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-    },
-    "transformIgnorePatterns": [
-      "[/\\\\]node_modules[/\\\\].+\\.(js|jsx)$"
-    ],
-    "moduleNameMapper": {
-      "^react-native$": "react-native-web"
-    },
-    "moduleFileExtensions": [
-      "web.js",
-      "js",
-      "json",
-      "web.jsx",
-      "jsx",
-      "node"
-    ]
-  },
-  "babel": {
-    "presets": [
-      "react-app"
-    ]
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
+    "test": "react-app-rewired test",
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
